// SETPPBO/ClientApp/app/main/abg/models/tanda-terima-subst/tanda-terima-subst.ts
export class TandaTerimaSubst {
    //list
    PermohonanId: string;
    PemohonId: string;
    NamaPemohon: string;
    NoSubSt: string;
    NoSuratPermintaanBantahan: string;
    NoSuratBantahan: string;
    NoSuratPermintaanSalinan: string;

    //form
    NoSengketa: string;
    RefJenisKetetapanId: number;
    NoTandaTerimaSubSt: string;
    TglTandaTerimaSubSt: Date;
    NoSuratPermintaanSubSt: string;
    TglSuratPermintaanSubSt: Date;
    TglTerimaAbgSubSt: Date;
    NamaTermohon: string;
    NamaUpTermohon: string;
    AlamatTermohon: string;
    KotaTermohon: string;
    
    // form edit permohonan
    NPWP: string;
    AlamatPemohon: string;
    KotaPemohon: string;
    NoSuratPermohonan: string;
    TglSuratPermohonan: Date;
    NoKep: string;
    TglKep: Date;
    JenisPajakId: number;
    JenisPajak: string;
    MasaPajakAwalTahun: number;
    MasaPajakAwalBulan: number;
    MasaPajakAkhirTahun: number;
    MasaPajakAkhirBulan: number;
    NoSkp: string;
    TglSkp: Date;
}